Refactor status effect scripts to remove unnecessary using directives and clean up code structure
- Removed unused `using System.Collections;` and `using System.Collections.Generic;` from various status effect scripts.
- Cleaned up whitespace and formatting in `BlindStatusEffect`, `HasteStatusEffect`, `KnockOutStatusEffect`, `PoisonStatusEffect`, `SilenceStatusEffect`, `SlowStatusEffect`, `StatusEffect`, and `StopStatusEffect`.
- Consolidated the `Awake` method in `Tile` class and moved the `Match` method to a private region for better organization.
- Changed `overrideFlags` in `TileTraversalOverride` to be readonly.
- Simplified the `Unit` class by merging `LoadData` and `SaveData` methods into the main class body.
- Updated shader files for better readability and consistency in formatting.
- Adjusted texture import settings for `Dirt.png` to improve compatibility and performance.
- Created a new meta file for shaders in the Resources folder.

+++ Assets/Scripts/Temp/TestLevelGrowth.cs
@@ -42,7 +42,7 @@ public class TestLevelGrowth : MonoBehaviour
             var actor = new GameObject(name);
             actor.AddComponent<Stats>();
             var rank = actor.AddComponent<Rank>();
-            rank.Init((int)Random.Range(1, 5));
+            rank.Init(Random.Range(1, 5));
             heroes.Add(actor);
         }
